Don't depend on ImageMagick or sips for rotate
Use Inkscape to do rotation. Currently a sort of a hack as to
do rotation without Inkscape causing aliasing we need to export
into double resolution and then export into the intended resolution.
This can probably be fixed in the future if Inkscape added support
for disabling bitmap oversampling per file or via command line.